Restructure markdown plugin (#18479)
