+++ src/language-markdown/index.js
@@ -1,8 +1,4 @@
-import printer from "./printer-markdown.js";
-
-export const printers = {
-  mdast: printer,
-};
 export { default as languages } from "./languages.evaluate.js";
 export { default as options } from "./options.js";
-export * as parsers from "./parser-markdown.js";
+export * as parsers from "./parsers.js";
+export * as printers from "./printers.js";
