+++ src/language-markdown/parse/index.js
@@ -1,0 +1,1 @@
+export { parseMarkdown, parseMdx } from "./parse.js";
