+++ src/language-markdown/parse/parse.js
@@ -2,9 +2,7 @@ import footnotes from "remark-footnotes";
 import remarkMath from "remark-math";
 import remarkParse from "remark-parse";
 import unified from "unified";
-import { locEnd, locStart } from "./loc.js";
 import { BLOCKS_REGEX, esSyntax } from "./mdx.js";
-import { hasIgnorePragma, hasPragma } from "./pragma.js";
 import frontMatter from "./unified-plugins/front-matter.js";
 import htmlToJsx from "./unified-plugins/html-to-jsx.js";
 import liquid from "./unified-plugins/liquid.js";
@@ -44,14 +42,7 @@ function createParse({ isMDX }) {
 
 function noop() {}
 
-const baseParser = {
-  astFormat: "mdast",
-  hasPragma,
-  hasIgnorePragma,
-  locStart,
-  locEnd,
-};
+const parseMarkdown = /* @__PURE__ */ createParse({ isMDX: false });
+const parseMdx = /* @__PURE__ */ createParse({ isMDX: true });
 
-export const markdown = { ...baseParser, parse: createParse({ isMDX: false }) };
-export const mdx = { ...baseParser, parse: createParse({ isMDX: true }) };
-export { markdown as remark };
+export { parseMarkdown, parseMdx };
