+++ src/language-markdown/print/index.js
@@ -1,0 +1,1 @@
+export { printMdast } from "./mdast.js";
