+++ src/language-markdown/print/whitespace.js
@@ -1,14 +1,14 @@
-import { hardline, line, softline } from "../document/index.js";
+import { hardline, line, softline } from "../../document/index.js";
 import {
   KIND_CJ_LETTER,
   KIND_CJK_PUNCTUATION,
   KIND_K_LETTER,
   KIND_NON_CJK,
-} from "./utilities.js";
+} from "../utilities.js";
 
 /**
- * @import {WordNode, WhitespaceValue, WordKind} from "./utilities.js"
- * @import AstPath from "../common/ast-path.js"
+ * @import {WordNode, WhitespaceValue, WordKind} from "../utilities.js"
+ * @import AstPath from "../../common/ast-path.js"
  * @typedef {"always" | "never" | "preserve"} ProseWrap
  * @typedef {{ next?: WordNode | null, previous?: WordNode | null }}
  * AdjacentNodes Nodes adjacent to a `whitespace` node. Are always of type
