+++ src/language-markdown/traverse/get-visitor-keys.js
@@ -1,4 +1,4 @@
-import createGetVisitorKeys from "../utilities/create-get-visitor-keys.js";
+import createGetVisitorKeys from "../../utilities/create-get-visitor-keys.js";
 import visitorKeys from "./visitor-keys.evaluate.js";
 
 const getVisitorKeys = createGetVisitorKeys(visitorKeys);
