+++ src/language-markdown/utilities.js
@@ -305,11 +305,16 @@ function getNthListSiblingIndex(node, parentNode) {
   }
 }
 
+function hasPrettierIgnore(path) {
+  return path.index > 0 && isPrettierIgnore(path.previous) === "next";
+}
+
 export {
   getFencedCodeBlockValue,
   getNthListSiblingIndex,
   getOrderedListItemInfo,
   hasGitDiffFriendlyOrderedList,
+  hasPrettierIgnore,
   INLINE_NODE_TYPES,
   INLINE_NODE_WRAPPER_TYPES,
   isAutolink,
